updated the log path for aliases
Each of the aliases has a path to write stdin and stderr to. The path
was hard coded to ox-hugo, but I think it makes more sense to write
this output to the dev parent directory. That way git won't see it as
a new file and the aliases can be used for projects other than ox-hugo
more easily.